@@ -7,33 +7,31 @@
     "start": "webpack-dev-server --open",
     "api": "babel-node --presets es2015 server.js"
   },
-  "author": "taqql.com",
+  "author": "resolff.com",
   "license": "ISC",
+
   "dependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "body-parser": "^1.17.2",
     "cors": "^2.8.4",
     "express": "^4.15.3",
     "jquery": "^3.2.1",
-    "node-static": "^0.7.9",
     "react": "^15.6.1",
     "react-bootstrap": "^0.31.0",
     "react-dom": "^15.6.1",
     "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "toastr": "^2.1.2",
-    "uuid": "^3.1.0"
+    "react-router-dom": "^4.1.1"
   },
   "devDependencies": {
+    "babel": "^6.23.0",
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.25.0",
     "babel-loader": "^7.1.1",
+    "babel-polyfill": "^6.23.0",
     "babel-plugin-import": "^1.2.1",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-latest": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "body-parser": "^1.17.2",
     "bootstrap": "^3.3.7",
     "css-hot-loader": "^1.3.0",
     "css-loader": "^0.28.4",
@@ -43,13 +41,16 @@
     "less": "^2.7.2",
     "less-loader": "^4.0.4",
     "node-sass": "^4.5.3",
+    "node-static": "^0.7.9",
     "react-hot-loader": "^3.0.0-beta.7",
     "sass-loader": "^6.0.6",
     "source-map-loader": "^0.2.1",
     "style-loader": "^0.18.2",
     "url-loader": "^0.5.9",
-    "webpack": "^2.2.0",
-    "webpack-dev-server": "^2.2.0"
+    "webpack": "^3.2.0",
+    "webpack-dev-server": "^2.5.1",
+    "toastr": "^2.1.2",
+    "uuid": "^3.1.0"
   },
   "theme": "./theme.js"
 }
