@@ -7,9 +7,9 @@ import {
     Link
 } from 'react-router-dom';
 
-export const Header = React.createClass({
+export class Header extends React.Component{
 
-    render: function () {
+    render() {
 
         return (
             <div>
@@ -65,4 +65,4 @@ export const Header = React.createClass({
             </div>
         )
     }
-});
+}
