@@ -1,28 +1,32 @@
 import React from "react";
 import {AutorPrevey} from "./AutorPrevey";
 import mass from "../mass.json";
+import {} from "react-router"
 
-export const InboxPage = React.createClass({
-    getInitialState(){
-        return{
+export class InboxPage extends React.Component{
+    constructor(props) {
+        super(props);
+        this.state = {
             mass
         };
-    },
+    }
 
     sowFullContent(massages){
-        const {mass}=this.state;
+        /*const {mass}=this.state;*/
         console.log(`/inbox/messages/${massages.id}`);
+        console.dir(this.props);
         this.props.history.push(`/inbox/messages/${massages.id}`);
 
-    },
-    render:function () {
-        const {mass}=this.state;
+    }
+
+    render(){
+       /* const {mass}=this.state;*/
         return(
             <div className="list-group">
                 {
                     mass.map(massages=>
                         <AutorPrevey
-                            onClick={this.sowFullContent.bind(null,massages)}
+                            onClick={this.sowFullContent.bind(this, massages)}
                             key = {massages.id}
                             id={massages.id}
                             shortInfo = {massages.shortInfo}
@@ -35,4 +39,4 @@ export const InboxPage = React.createClass({
         )
     }
 
-});
+}
